Fixed gift records price and error msg

See merge request taya/backend/web-api!37

+++ serializer/gift_record.go
@@ -10,13 +10,13 @@ type GiftRecord struct {
 	ID     string `json:"id"`
 	UserId int64  `json:"user_id"`
 	// GiftId       int64     `json:"gift_id"`
-	GiftName      string `json:"gift_name"`
-	Quantity      int    `json:"quantity"`
-	TotalPrice    int64  `json:"total_price"`
-	LiveStreamId  int64  `json:"live_stream_id"`
-	GiftTime      int64  `json:"gift_time"`
-	TransactionID string `json:"transaction_id"`
-	StreamerName  string `json:"streamer_name"`
+	GiftName      string  `json:"gift_name"`
+	Quantity      int     `json:"quantity"`
+	TotalPrice    float64 `json:"total_price"`
+	LiveStreamId  int64   `json:"live_stream_id"`
+	GiftTime      int64   `json:"gift_time"`
+	TransactionID string  `json:"transaction_id"`
+	StreamerName  string  `json:"streamer_name"`
 }
 
 type PaginatedGiftRecord struct {
@@ -30,8 +30,8 @@ func BuildPaginatedGiftRecord(a []models.GiftRecord, total, amount, win int64) (
 
 	b = PaginatedGiftRecord{
 		TotalCount:  total,
-		TotalAmount: float64(amount) / 100,
-		TotalWin:    float64(win) / 100,
+		TotalAmount: float64(amount) / float64(100),
+		TotalWin:    float64(win) / float64(100),
 	}
 
 	for _, giftRecord := range a {
@@ -43,7 +43,7 @@ func BuildPaginatedGiftRecord(a []models.GiftRecord, total, amount, win int64) (
 			ID:     uniqueId,
 			UserId: giftRecord.UserId,
 			// GiftId:       giftRecord.GiftId,
-			TotalPrice:    giftRecord.TotalPrice / 100,
+			TotalPrice:    float64(giftRecord.TotalPrice) / float64(100),
 			Quantity:      giftRecord.Quantity,
 			LiveStreamId:  giftRecord.LiveStreamId,
 			GiftTime:      giftRecord.CreatedAt.Unix(),
